Add dealer-locator, header blocks
Automated migration updates generated by Experience Modernization Agent.

Changed files (3):
- blocks/dealer-locator/dealer-locator.js
- blocks/header/header.js
- scripts/helpers.js
